pyproject.toml
[tool.poetry]
name = "newsgroups-qwen-embed"
version = "0.1.0"
description = "Demo project for exploring QWEN-3 embeddings with newsgroups dataset"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
aiohttp = "^3.9.0"
aiolimiter = "^1.1.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"
tenacity = "^8.2.0"
scikit-learn = "^1.3.0"
umap-learn = "^0.5.0"
plotly = "^5.0.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
notebook = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
